Modify AuthVC code
